Join router and handler modules
