+++ src/handler/routing.rs
@@ -1,174 +1,4 @@
-//!Routers stores request handlers, using an HTTP method and a path as keys.
-//!
-//!# Building Routers
-//!
-//!Rustful provides a tree structured all-round router called `DefaultRouter`,
-//!but any other type of router can be used, as long as it implements the
-//!`Router` trait. This will also make it possible to initialize it using the
-//![`insert_routes!`][insert_routes] macro:
-//!
-//!```
-//!#[macro_use]
-//!extern crate rustful;
-//!use rustful::DefaultRouter;
-//!# use rustful::{Handler, Context, Response};
-//!
-//!# struct ExampleHandler;
-//!# impl Handler for ExampleHandler {
-//!#     fn handle(&self, _: Context, _: Response){}
-//!# }
-//!# fn main() {
-//!# let about_us = ExampleHandler;
-//!# let show_user = ExampleHandler;
-//!# let list_users = ExampleHandler;
-//!# let show_product = ExampleHandler;
-//!# let list_products = ExampleHandler;
-//!# let show_error = ExampleHandler;
-//!# let show_welcome = ExampleHandler;
-//!let router = insert_routes! {
-//!    DefaultRouter::<ExampleHandler>::new() => {
-//!        Get: show_welcome,
-//!        "about" => Get: about_us,
-//!        "users" => {
-//!            Get: list_users,
-//!            ":id" => Get: show_user
-//!        },
-//!        "products" => {
-//!            Get: list_products,
-//!            ":id" => Get: show_product
-//!        },
-//!        "*" => Get: show_error
-//!    }
-//!};
-//!# }
-//!```
-//!
-//!This macro creates the same structure as the example below, but it allows
-//!tree structures to be defined without the need to write the same paths
-//!multiple times. This can be useful to lower the risk of typing errors,
-//!among other things.
-//!
-//!Routes may also be added using the insert method, like this:
-//!
-//!```
-//!extern crate rustful;
-//!use rustful::Method::Get;
-//!use rustful::{Insert, DefaultRouter};
-//!# use rustful::{Handler, Context, Response};
-//!
-//!# struct ExampleHandler;
-//!# impl Handler for ExampleHandler {
-//!#     fn handle(&self, _: Context, _: Response){}
-//!# }
-//!# fn main() {
-//!# let about_us = ExampleHandler;
-//!# let show_user = ExampleHandler;
-//!# let list_users = ExampleHandler;
-//!# let show_product = ExampleHandler;
-//!# let list_products = ExampleHandler;
-//!# let show_error = ExampleHandler;
-//!# let show_welcome = ExampleHandler;
-//!let mut router = DefaultRouter::<ExampleHandler>::new();
-//!
-//!router.insert(Get, "/", show_welcome);
-//!router.insert(Get, "/about", about_us);
-//!router.insert(Get, "/users", list_users);
-//!router.insert(Get, "/users/:id", show_user);
-//!router.insert(Get, "/products", list_products);
-//!router.insert(Get, "/products/:id", show_product);
-//!router.insert(Get, "/*", show_error);
-//!# }
-//!```
-//!
-//![insert_routes]: ../macro.insert_routes!.html
-//!
-//!#Variables
-//!
-//!Routes may contain variables, that are useful for capturing parts of the
-//!requested path as input to the handler. The syntax for a variable is simply
-//!an indicator character (`:` or `*`) followed by a label. Variables without
-//!labels are also valid, but their values will be discarded.
-//!
-//!##Variable Segments (:label)
-//!
-//!A variable segment will match a single arbitrary segment. They are probably
-//!the most commonly used variables and may, for example, be used to select a
-//!blog post: `"posts/:year/:month/:day/:title_slug"`.
-//!
-//!```text
-//!pattern = "a/:v/b"
-//!"a/c/b" -> v = "c"
-//!"a/c/d/b" -> no match
-//!"a/b" -> no match
-//!"a/c/b/d" -> no match
-//!```
-//!
-//!##Variable Sequences (*label)
-//!
-//!A variable sequence is similar to a variable segment, but with the
-//!difference that it may consume multiple segments until the rest of the path
-//!gives a match. An example use case is a route for downloadable files that
-//!may be arranged in arbitrary directories: `"downloads/*file_path"`.
-//!
-//!```text
-//!pattern = "a/*v/b"
-//!"a/c/b" -> v = "c"
-//!"a/c/d/b" -> v = "c/d"
-//!"a/b" -> no match
-//!"a/c/b/d" -> no match
-//!```
-//!
-//!```text
-//!pattern = "a/b/*v"
-//!"a/b/c" -> v = "c"
-//!"a/b/c/d" -> v = "c/d"
-//!"a/b" -> no match
-//!```
-//!
-//!# Router Composition
-//!
-//!The default router is actually a composition of three routers:
-//![`TreeRouter`][tree_router], [`MethodRouter`][method_router] and
-//![`Variables`][variables]. They come together as the type
-//!`TreeRouter<MethodRouter<Variables<_>>>`, but the `TreeRouter` assumes that
-//!this is most probably what you want, so this is what `TreeRouter::new()`
-//!gives you. No need to write it all out in most of the cases.
-//!
-//!There may, however, be cases where you want something else. What if you
-//!don't care about the HTTP method? Maybe your handler takes care of that
-//!internally. Sure, no problem:
-//!
-//!```
-//!use rustful::router::{Variables, TreeRouter};
-//!
-//!let my_router = TreeRouter::<Option<Variables<_>>>::default();
-//!# let _r: TreeRouter<Option<Variables<ExampleHandler>>> = my_router;
-//!# struct ExampleHandler;
-//!# impl rustful::Handler for ExampleHandler {
-//!#     fn handle(&self, _: rustful::Context, _: rustful::Response){}
-//!# }
-//!```
-//!
-//!And what about those route variables? Not using them at all? Well, just
-//!remove them too, if you don't want them:
-//!
-//!```
-//!use rustful::router::TreeRouter;
-//!
-//!let my_router = TreeRouter::<Option<_>>::default();
-//!# let _r: TreeRouter<Option<ExampleHandler>> = my_router;
-//!# struct ExampleHandler;
-//!# impl rustful::Handler for ExampleHandler {
-//!#     fn handle(&self, _: rustful::Context, _: rustful::Response){}
-//!# }
-//!```
-//!
-//!You can simply recombine and reorder the router types however you want, or
-//!why not make your own router? Just implement the `Router` trait.
-//!
-//![tree_router]: struct.TreeRouter.html
-//![method_router]: struct.MethodRouter.html
-//![variables]: struct.Variables.html
+//!Routing related traits and types.
 
 use std::collections::HashMap;
 use std::iter::{Iterator, FlatMap, Peekable};
@@ -179,20 +9,6 @@ use hyper::method::Method;
 
 use handler::Handler;
 use context::MaybeUtf8Owned;
-
-pub use self::tree_router::TreeRouter;
-pub use self::method_router::MethodRouter;
-pub use self::variables::Variables;
-
-mod tree_router;
-mod method_router;
-mod variables;
-
-///Alias for `TreeRouter<MethodRouter<Variables<T>>>`.
-///
-///This is probably the most common composition, which will select handlers
-///based on path and then HTTP method, and collect any variables on the way.
-pub type DefaultRouter<T> = TreeRouter<MethodRouter<Variables<T>>>;
 
 ///A common trait for building routers.
 ///
@@ -281,7 +97,7 @@ pub trait Route<'a> {
     ///None for a root path (`/`).
     ///
     ///```rust
-    ///# use rustful::router::Route;
+    ///# use rustful::handler::routing::Route;
     ///let root = "/";
     ///assert_eq!(root.segments().next(), None);
     ///
